Add Feature To Centreline Partitioner
Added end points as an optional parameter for the Centreline Partitioner.
The user can specify a number of point Ids; the first decides the starting
point of the partitioner and the rest are treated as new endpoints. These
end points only determine what cells are created, all points from the
original vtkPolyData are still present.