Add Filter To Subdivide Meshs
Created two filters, one that subdivides a given cell into as many
rows and columns specified as parameters, and one that uses this
to build smaller quads for every quad in a mesh. This second filter,
vtkSubdivideMeshDynamic, accepts a width and length that is to be the
size of new quads within each original quad and calculates roughly how
many can be present (as the original quads can vary in size). For large
data sets this filter is quite expensive. testSubdivideMeshStatic
ignores the difference in size between original quads and simply builds
the same number of new quads in each original. It is not a filter
currently, just a test file.