Add Pipeline Tests And Modified Synthetic Centreline Generator
Created a full pipeline test that builds a quad mesh, a quad mesh with
skip segments and caps, and EC and SMC quad meshes. A very similar
test that does the same around a bifurcation is also included.

SynthCentrelineGenerator has been modified to provide a smoother
decrement of radii scalar values.