Minor Edits, Bug Detected
When the code is figuring it out whether a point is a part of a
bifurcation, the test with number of shared cells greater than
two fails, because non-zero segment starting points participate
it four cells each. Perhaps the way to go here is to store
bifurcation ids in a polyvertex object.