Revise the Code to Generate Non-Bifurcating Segments
The code in the src/vtkScalarRadiiToVectorsFilter.cxx was revised to allow straight segments mesh generation.

Clean up the directories under data and tmpData to keep track of the *.python files for generating ATP files and legacy format files.

Remove *.vtk centreline files because we don't need to store generated data.