Convert c216 mesh to Python 3; Remove obsolete source files for ATP legacy mesh generation.
Correct print calls and such to make the code work in Python 3 for just one mesh, c216. Other
meshes still won't work in Python 3, but the changes should be trivial.

The source file DumpATPMeshToLegacyForhmat.py has been removed, but launcher scripts pointing
to the file have been left behind, which are removed in this commit.